Make BearerToken equality comparison more efficient
